% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CD2}
\alias{CD2}
\title{Centered L2 discrepancy using C-->C++
The vector x is read in rowwise rather than columnwise.
To Obtain the correct values, you should transpose the matrix}
\usage{
CD2(x)
}
\arguments{
\item{x}{The design matrix D}
}
\description{
Centered L2 discrepancy using C-->C++
The vector x is read in rowwise rather than columnwise.
To Obtain the correct values, you should transpose the matrix
}
