% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{construct1}
\alias{construct1}
\title{Construct N x (N - 1) good lattice points}
\usage{
construct1(n)
}
\arguments{
\item{n}{Number of rows for the final design. SHOULD BE PRIME}
}
\value{
design d with n rows, n levels, n - 1 columns
}
\description{
Construct N x (N - 1) good lattice points
}
