% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{d2}
\alias{d2}
\title{The L2 distance of design matrix x. min{d_p(x_i,x_j), 1<=i<j<=n}
equation 2.1}
\usage{
d2(x)
}
\arguments{
\item{x}{The Design matrix x}
}
\description{
The L2 distance of design matrix x. min{d_p(x_i,x_j), 1<=i<j<=n}
equation 2.1
}
