% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{phi efficiency}
\alias{phi efficiency}
\alias{phiEff}
\title{Computes the phi efficiency.}
\usage{
phiEff(x, k = 2L)
}
\arguments{
\item{x}{Design matrix x}
}
\description{
Computes the phi efficiency.
}
