% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{phi}
\alias{phi}
\title{Compute phi value for design D . Check equation 3.1 Uniform Projection
Designs paper combn + mean.}
\usage{
phi(x, k = 2L)
}
\arguments{
\item{x}{input design D}

\item{k}{number of dimensions to be projected onto}
}
\description{
Compute phi value for design D . Check equation 3.1 Uniform Projection
Designs paper combn + mean.
}
