% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{phi2}
\alias{phi2}
\title{Used Combn + reduce function.}
\usage{
phi2(x, k)
}
\arguments{
\item{x}{the design matrix x}

\item{k}{the number of dimensions to be projected onto.}
}
\description{
Used Combn + reduce function.
}
