% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{phi2D}
\alias{phi2D}
\title{Compute phi value for design D . Check equation 3.1 Uniform Projection
Designs paper}
\usage{
phi2D(x, s)
}
\arguments{
\item{x}{input design D}

\item{s}{The level for the design matrix D}
}
\description{
Assumes we are only interested in 2d projection
}
\examples{
Uni::phi2D(construct(7), 7);
}
