% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{phi_smallest}
\alias{phi_smallest}
\title{eturn the design with smallest phi among those projected into K dimensions}
\usage{
phi_smallest(x, k = 2L)
}
\arguments{
\item{x}{the design matrix d}

\item{k}{the number of dimensions to be projected onto.}
}
\description{
eturn the design with smallest phi among those projected into K dimensions
}
