% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rho_ave}
\alias{rho_ave}
\title{average of the absolute correlations between columns
Design is orthogonal if rho_ave(D) == 0}
\usage{
rho_ave(x)
}
\arguments{
\item{x}{The design matrix D}
}
\description{
average of the absolute correlations between columns
Design is orthogonal if rho_ave(D) == 0
}
