% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WilliamInv}
\alias{WilliamInv}
\title{Inverse William transformation}
\usage{
WilliamInv(x, levels)
}
\arguments{
\item{x}{The Design matrix x}

\item{levels}{The number of levels in the design}
}
\description{
Inverse William transformation
}
